+++ pro-common/src/main/java/dr/sbs/common/CommonPage.java
@@ -9,7 +9,7 @@ public class CommonPage<T> {
   private Integer pageNum;
   private Integer pageSize;
   private Integer pages;
-  private long total;
+  private Integer total;
   private List<T> list;
 
   public static <T> CommonPage<T> toPage(IPage<T> p) {
@@ -18,7 +18,7 @@ public class CommonPage<T> {
     result.setPages((int) p.getPages());
     result.setPageNum((int) p.getCurrent());
     result.setPageSize((int) p.getSize());
-    result.setTotal(p.getTotal());
+    result.setTotal((int) p.getTotal());
     result.setList(p.getRecords());
 
     return result;
